Add :discard support for :lexical-preservation t, and show that it's broken
Discard nodes were currently not being preserved, this changes handles
reductions of type :discard, but also adds a test case that demonstrates this
breaks down in the face of whitespace (or other discard nodes).

The reason is that the parser only considers the top two elements on the stack,
but in the case of :lexical-preservation, there could be several :whitespace and
:comment nodes that need to be taken into account as well.

We'll have the same problem when dealing with metadata.