Fix parsing of tags/discard with :lexical-preservation
When nodes on the stack can be "empty" (i.e. whitespace, comments), then extra
care needs to be taken to prevent e.g. a #_ from discarding the whitespace node
that follows it in "#_ 1 2 3".

This introduces parseclj--take-token and parseclj--take-value, which should also
come in handy to implement three-element reduction rule for metadata literals.